Revert "Feature/drawer names"
