--- lib/views/archive.dart
@@ -1,13 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:ddsr/functs/drawer.dart';
-
-class ArchivePage extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(title:  new Text('asdf')),
-      body: SafeArea(child: Text('second page')),
-      drawer: BaseDrawer(),
-    );
-  }
-}
